Remove lots of unused things.
