+++ entwine/io/zstandard.cpp
@@ -10,9 +10,9 @@
 
 #include <entwine/io/zstandard.hpp>
 
+/*
 #include <pdal/compression/ZstdCompression.hpp>
 
-/*
 namespace entwine
 {
