+++ entwine/types/CMakeLists.txt
@@ -7,7 +7,6 @@ set(
     "${BASE}/file-info.cpp"
     "${BASE}/files.cpp"
     "${BASE}/metadata.cpp"
-    "${BASE}/pooled-point-table.cpp"
     "${BASE}/subset.cpp"
 )
 
@@ -23,17 +22,14 @@ set(
     "${BASE}/fixed-point-layout.hpp"
     "${BASE}/key.hpp"
     "${BASE}/metadata.hpp"
-    "${BASE}/outer-scope.hpp"
     "${BASE}/point.hpp"
-    "${BASE}/point-pool.hpp"
-    "${BASE}/pooled-point-table.hpp"
     "${BASE}/reprojection.hpp"
     "${BASE}/schema.hpp"
     "${BASE}/stats.hpp"
     "${BASE}/subset.hpp"
-    "${BASE}/tube.hpp"
     "${BASE}/vector-point-table.hpp"
     "${BASE}/version.hpp"
+    "${BASE}/voxel.hpp"
 )
 
 install(FILES ${HEADERS} DESTINATION include/entwine/${MODULE})
