+++ entwine/types/defs.hpp.in
@@ -14,7 +14,8 @@
 #include <string>
 #include <vector>
 
-#include <entwine/third/bigint/little-big-int.hpp>
+#include <pdal/Dimension.hpp>
+
 #include <entwine/types/bounds.hpp>
 #include <entwine/types/version.hpp>
 
@@ -30,15 +31,14 @@ inline Version currentVersion()
     return Version(ENTWINE_VERSION_STRING);
 }
 
-using Id = BigUint;
-
 using Origin = uint64_t;
 using OriginList = std::vector<Origin>;
 static constexpr Origin invalidOrigin = std::numeric_limits<Origin>::max();
 
-using TileFunction = std::function<void(pdal::PointView& view, Bounds bounds)>;
-
 using Paths = std::vector<std::string>;
+
+using DimId = pdal::Dimension::Id;
+using DimType = pdal::Dimension::Type;
 
 } // namespace entwine
